Create word_match_fiesta.py
This is my version of the Word Match Fiesta from VS Code Education.

It is a language learning game where the player must match an English word with the corresponding Spanish word. Each time they get an answer correct that result is removed from the matrix until there are no words left.